feat(ioriver-go): New release
